requirements.txt
streamlit==1.52.0
llama-index==0.11.10
llama-index-core==0.11.10
llama-index-llms-google-genai==0.2.0
llama-index-embeddings-huggingface==0.2.0
google-generativeai==0.8.3
transformers==4.36.2
torch
python-dotenv
pandas
numpy
tqdm
